Restructuring code before merge
