+++ adventure/adventure.py
@@ -1,0 +1,2 @@
+import program
+program.main()
