+++ adventure/data/creatures.py
