+++ adventure/program/__init__.py
@@ -1,0 +1,3 @@
+from program.parser import parse
+from program.utils import describe_room
+from program.main import main
