+++ adventure/program/parser.py
@@ -1,40 +1,39 @@
-import Commands
+import program.commands as command
 
+def parse(input):
 
-def parse(command):
-
-    match list(filter(lambda word : word not in ("a", "an", "the"), command.split())):
+    match list(filter(lambda word : word not in ("a", "an", "the"), input.split())):
     # split command string and remove articles, before matching cases, output is list of strings
 
         case ["n"] | ["north"]:
-            Commands.go("north")
+            command.go("north")
         
         case ["e"] | ["east"]:
-            Commands.go("east")
+            command.go("east")
 
         case ["s"] | ["south"]:
-            Commands.go("south")
+            command.go("south")
         
         case ["w"] | ["west"]:
-            Commands.go("west")
+            command.go("west")
 
         case ["go", "to", direction] | ["go", direction]:
-            Commands.go(direction)
+            command.go(direction)
        
         case ["look", "at", *item] | ["examine", *item] | ["ex", *item]:                # if you use *item, it will be a list (of strings), if just item, it will be a string
-            Commands.examine(item)
+            command.examine(item)
 
         case ["l"] | ["look"] | ["look", "around"]:
-            Commands.look_around()
+            command.look_around()
 
         case ["take", *item] | ["get", *item]:
-            Commands.take(item)
+            command.take(item)
 
         case ["drop", *item]:
-            Commands.drop(item)
+            command.drop(item)
 
         case ["i"] | ["inventory"] | ["check", "pockets"] | ["check", "pocket"]:
-            Commands.inventory()
+            command.inventory()
 
         case ["quit"]:
             print("\nQuitters!\n")
