--- object_oriented_programming_tests/Items_test.py
@@ -1,29 +1,0 @@
-
-class Item:
-    name = None
-    #...
-
-class Openable:
-    is_opened = False
-
-    def open(self):
-        self.is_opened = True
-       
-    def close(self):        
-        self.is_opened = False
-
-class Container:
-    items_inside = []
-
-class Chest(Item, Openable, Container):
-
-# --------------------
-
-item = Chest()
-
-if item is Openable and item.is_opened:
-    
-    ...
-
-
-
