--- other_tests/main.py
@@ -1,10 +1,0 @@
-
-from parser import parse
-
-raw_command = input("What do? ")
-
-command, parameters = parse(raw_command)
-
-print("Do:", command)
-print("Params:", parameters)
-
